Basic Multiplayer Fundamentals
Can host a server and join the host by using their IP as of now. Sends players to the lobby where they can see each other moving around. Once the minimum amount of players have joined, it'll send them to the next scene(which will eventually be the first minigame).