River Ruckus Adaptations
Adapted River Ruckus's scene hierarchy to fit the pattern of the rest of the minigames. Lowered the terrain so the player respawns at the water level. Adjusted the wisp spawn area to only be on one side of the terrain as intended. Replaced Veiled Threat with River Ruckus in the GameManager.