Bugfix: packet loss
The player sync script would cause a lot of packet loss so I'm disabling and replacing with the unreliable network transform instead. Will need more testing for desync. At this point, it would be better to have player movement desync over the packet loss anyways. Made some changes to the item controller to make it more client based to be more fluent and in case of packet loss.