Bugfix: remove remoteTimeStamp calculation
Calculating removeTimeStamp into scripts is nice because it makes sure everything happens at the same time for all clients but the issue is other stuff can't predict the future using remoteTimeStamp like the player which causes issues when they interact. This is why you would sometimes see the player is the ground or floating when on a platform moving vertically. Removing this as of now. Includes the timer, MoveObjectOverTime, and ConstantRotation. This was already removed from the jumping fish previously.